# Improve contains check done by FastRegexMatcher