@bboreham 2024-06-04
To explain my thinking: previously the test used 18 strings, total about 456 bytes.  This is small enough that most branches could fit in the prediction cache of the cpu, and adding one more branch could tip that change.  It's unrealistically small compared to any real-world Prometheus.

Increasing to 75,000 strings and 3.5MB makes it unlikely the CPU can predict every branch, and pushes out from L1 to L2/L3 cache.